tests/dog-api-tests/requirements.txt
pytest==8.3.4
requests==2.32.3
pytest-html==4.1.1
pytest-ordering==0.6
python-dotenv==1.0.1
